requirement.txt
pyfiglet
pymongo
os
python-dotenv
base64
requests
coloroma